+++ far/fn.hpp
@@ -1069,7 +1069,7 @@ string &Add_PATHEXT(string &strDest);
 extern "C" {
 #endif
 
-void __cdecl qsortex(char *base, unsigned int nel, unsigned int width,
+void __cdecl qsortex(char *base, size_t nel, size_t width,
             int (__cdecl *comp_fp)(const void *, const void *,void*), void *user);
 
 char * __cdecl farmktemp(char *temp);
